% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_types.R
\docType{class}
\name{ArrayType}
\alias{ArrayType}
\alias{typesys::ArrayType-class}
\title{Array Type}
\usage{
ArrayType(type, dimension = list(UnknownValue()))
}
\description{
A dimensioned container for homogeneous types. This is a superclass for
matrices.
}
