% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_types.R
\docType{class}
\name{Call}
\alias{Call}
\alias{typesys::Call-class}
\title{Call}
\usage{
Call(func, ...)
}
\description{
A call to a function.
}
\section{Slots}{

\describe{
\item{\code{func}}{The name of the function to call}

\item{\code{args}}{A list of argument types}
}}

