% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{class}
\name{CompositeType-class}
\alias{CompositeType-class}
\title{Composite Type}
\description{
A container for other (possibly composite) types. This is a superclass for
structures, lists, arrays, and matrices.
}
\section{Slots}{

\describe{
\item{\code{types}}{A named list of contained types.}
}}

