% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_type.R
\name{ConditionalType}
\alias{ConditionalType}
\title{Conditional Type}
\usage{
ConditionalType(handler = function(args) { })
}
\description{
A type dependent on a condition. See UnionType.
}
\details{
ConditionalType uses LIFO condition checking, so conditions should
be mutually exclusive or added in increasing order of granularity.
Internally, conditions are stored as an if-else statement.
}

