% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint.R
\docType{class}
\name{Constraint-class}
\alias{Constraint-class}
\title{Constraint}
\description{
Superclass for all constraints on terms.
}
\details{
Constraints are pairs of terms with a relation between them. Constraint
subclasses represent specific relations.
}
