% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csolve.R
\docType{methods}
\name{csolve}
\alias{csolve}
\alias{csolve,list-method}
\title{Solve a List of Constraints}
\usage{
csolve(x, ...)

\S4method{csolve}{list}(x, counter, max_deferred = 1000L, ...)
}
\arguments{
\item{x}{An object that contains a list of constraints to solve.}

\item{...}{Additional arguments to methods.}

\item{counter}{(Counter) A Counter for generating new variables.}

\item{max_deferred}{(integer) Maximum number of times the solver can defer
constraints before concluding the system has no solutions.}
}
\value{
The solution to the supplied constraints as a Substitution object.
}
\description{
This function solves a list of constraints.
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: Solve a list of constraints.
}}

\references{
Heeren, Hage, and Swierstra (2002). Generalizing Hindley-Milner Type
Inference Algorithms.
}
