% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitution.R
\name{do_substitution}
\alias{do_substitution}
\title{Apply a Substitution}
\usage{
do_substitution(term, sub)
}
\arguments{
\item{term}{A Term or Substitution object.}

\item{sub}{A Substitution object to apply.}
}
\value{
An object of the same class as \code{term}.
}
\description{
This generic function applies a Substitution to a Term or to another
Substitution (composition).
}
\details{
Substitutions are callable, and this is the generic they call. For clarity,
calling a Substitution directly on a term is preferable to calling this
function.
}
