% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{element_type}
\alias{element_type}
\title{Extract Contained Type}
\usage{
element_type(self)
}
\description{
For an atomic type or a container type with homogeneous elements, return the
underlying element type. For a continer type with heterogeneous elements,
throw an error.
}
