% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_to_type.R
\name{formula_to_type}
\alias{formula_to_type}
\title{Convert a Formula to a Type Expression}
\usage{
formula_to_type(x)
}
\arguments{
\item{x}{(formula) The formula to convert to a type expression.}
}
\description{
This function makes it possible to write type expressions concisely using a
formula-based language.
}
\details{
Literal types have the same names and capitalization as their S4 classes,
but with \code{typesys::} and \code{Type} excluded. So
\code{typesys::IntegerType} is just written as \code{Integer}. Type
variables must be written in lower case.
}
