% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instantiate.R
\name{instantiate}
\alias{instantiate}
\title{Instantiate Variables in a Term or Constraint}
\usage{
instantiate(term, ignore = list(), counter = NULL,
  helper = InstantiateHelper$new(counter))
}
\arguments{
\item{term}{Term or constraint to instantiate.}

\item{ignore}{List of Variables to ignore when replacing.}

\item{counter}{Counter for generating new variables.}

\item{helper}{Helper object for the function.}
}
\description{
This function instantiates variables in the given term or constraint,
replacing them with new variables.
}
