% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{class}
\name{Join}
\alias{Join}
\alias{typesys::Join-class}
\title{Join}
\usage{
Join(..., simplify = TRUE)
}
\description{
Take the join (least upper bound) of a list of types. This differs from a
Union in that it always returns a named type.
}
