% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_type.R
\name{pushCondition}
\alias{pushCondition}
\title{Push Condition}
\usage{
pushCondition(self, condition, if_type, else_type)
}
\description{
Push a condition onto the stack of conditions in a ConditionalType.
}

