% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_type.R
\name{pushCondition_q}
\alias{pushCondition_q}
\title{Push Quoted Condition}
\usage{
pushCondition_q(self, condition, if_type, else_type)
}
\description{
Push a quoted condition onto the stack of conditions in a ConditionalType.
}

