% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify.R
\name{quantify}
\alias{quantify}
\title{Quantify a Type Variable}
\usage{
quantify(type, tenv = NULL)
}
\arguments{
\item{type}{(Type) The type expression to quantify.}

\item{tenv}{(TypeEnvironment) The type environment to check for bound type
variables. If NULL, all type variables are quantified.}
}
\description{
This function quantifies all type variables in the supplied type expression
that aren't bound in the supplied type environment.
}
