% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scope.R
\name{Scope}
\alias{Scope}
\title{Scope}
\usage{
Scope(parent = NULL)
}
\description{
This R6 class represents a scope within a program. In particular, a scope
can have a parent scope, a list of contained names, and a return value (to
model functions).
}
