% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{simplify}
\alias{simplify}
\title{Simplify a Join}
\usage{
simplify(join)
}
\arguments{
\item{join}{(Join) The join to simplify.}
}
\description{
This function simplifies a join so that it does not contain any nested joins
and each element is unique.
}
