% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve}
\alias{solve}
\title{Solve a List of Constraints}
\usage{
solve(constraints, counter)
}
\arguments{
\item{constraints}{List of Constraint objects to solve.}

\item{counter}{Counter for generating new variables.}
}
\value{
The solution to the supplied constraints as a Substitution object.
}
\description{
This function solves a list of constraints.
}
\references{
Heeren, Hage, and Swierstra (2002). Generalizing Hindley-Milner Type
Inference Algorithms.
}
