% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{class}
\name{Type-class}
\alias{Type-class}
\title{Type}
\description{
Parent class for all types.
}
\section{Slots}{

\describe{
\item{\code{value}}{A value, for constant folding}

\item{\code{contexts}}{(character) Contextual information about the type.}
}}

