% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_environment.R
\docType{data}
\name{TypeEnvironment}
\alias{TypeEnvironment}
\title{Type Environment}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TypeEnvironment
}
\description{
Type Environment
}
\keyword{datasets}
