% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_types.R
\docType{class}
\name{UndefinedType}
\alias{UndefinedType}
\alias{typesys::UndefinedType-class}
\title{Undefined Type}
\usage{
UndefinedType()
}
\description{
"Bottom" type in the type lattice, which indicates a variable which is
nonexistent or has not yet been initialized.
}
