% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unify.R
\name{unify}
\alias{unify}
\title{Robinson's Unification Algorithm}
\usage{
unify(x, y, sub = Substitution())
}
\description{
Given two types, this method returns a substitution that can be applied to
both to make them equal. If no such substitution exists, the method throws
an error.
}
