% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{class}
\name{Union}
\alias{Union}
\alias{Union-class}
\title{Union}
\usage{
Union(...)
}
\description{
A union of two or more types.
}
\section{Slots}{

\describe{
\item{\code{types}}{A list of types.}
}}

