% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{class}
\name{UnionType}
\alias{UnionType}
\alias{UnionType-class}
\title{Union Type}
\usage{
UnionType(...)
}
\description{
A union of two or more types.
}
\section{Slots}{

\describe{
\item{\code{types}}{A list of possible types.}
}}

