% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_types.R
\docType{class}
\name{UnknownType}
\alias{UnknownType}
\alias{UnknownType-class}
\title{Unknown Type}
\usage{
UnknownType()
}
\description{
"Top" type in type lattice, which indicates a variable whose type is
unknown.
}

