% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_terms.R
\name{vars}
\alias{vars}
\title{Get Variables}
\usage{
vars(term, generic = FALSE)
}
\arguments{
\item{term}{(Term|Constraint) The term to search for variables.}

\item{generic}{(logical) Should generic variables be included in the list?}
}
\value{
A list of Variable objects.
}
\description{
This function returns a list of all variables in a term or constrain.
}
\details{
In an ImplicitInstance constraint, some variables in the right-hand term may
be generic rather than monomorphic. Generic variables are replaced with new
variables when a constraint is solved, whereas monomorphic variables are
not. The \code{generic} parameter controls whether generic variable are included
in returned the list of variables.
}
