% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{class}
\name{VectorType}
\alias{VectorType}
\alias{typesys::VectorType-class}
\title{Vector Type}
\usage{
VectorType(type, dimension = list())
}
\description{
A dimensioned container for homogeneous types. Unlike \code{ArrayType}, this
class represents a container that can grow dynamically.
}
