% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{class}
\name{CompositeType-class}
\alias{CompositeType-class}
\alias{length,CompositeType-method}
\title{Composite Type}
\usage{
\S4method{length}{CompositeType}(x)
}
\description{
A container for other (possibly composite) types. This is a superclass for
structures, lists, arrays, and matrices.
}
\section{Methods (by generic)}{
\itemize{
\item \code{length}: Length of the type.
}}
\section{Slots}{

\describe{
\item{\code{types}}{A named list of contained types.}
}}

