% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\docType{methods}
\name{names,typesys::Variable-method}
\alias{names,typesys::Variable-method}
\alias{names,typesys::Constant-method}
\alias{names,typesys::Composite-method}
\alias{names<-,typesys::Variable-method}
\title{Variable Names}
\usage{
\S4method{names}{`typesys::Variable`}(x)

\S4method{names}{`typesys::Constant`}(x)

\S4method{names}{`typesys::Composite`}(x)

\S4method{names}{`typesys::Variable`}(x) <- value
}
\arguments{
\item{x}{A Term object.}
}
\value{
A character vector of variable names.
}
\description{
Functions to get the variable names in a Term. Names can only be set on a
Variable.
}
