% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve}
\alias{solve}
\title{Solve a List of Constraints}
\usage{
solve(constraints, counter, max_deferred = 1000L)
}
\arguments{
\item{constraints}{(list) A list of Constraint objects to solve.}

\item{counter}{(Counter) A Counter for generating new variables.}

\item{max_deferred}{(integer) Maximum number of times the solver can defer
constraints before concluding the system has no solutions.}
}
\value{
The solution to the supplied constraints as a Substitution object.
}
\description{
This function solves a list of constraints.
}
\references{
Heeren, Hage, and Swierstra (2002). Generalizing Hindley-Milner Type
Inference Algorithms.
}
